Move categoricalembedding and continuousembedding plots into sharedPlots
