+++ src/__test__/components/sharedPlots/ContinuousEmbeddingPlot.test.jsx
@@ -7,12 +7,12 @@ import thunk from 'redux-thunk';
 import { Provider } from 'react-redux';
 import { Vega } from 'react-vega';
 
-import ContinuousEmbeddingPlot from '../../../../../../../pages/experiments/[experimentId]/data-processing/configure-embedding/components/ContinuousEmbeddingPlot';
-import { initialEmbeddingState } from '../../../../../../../redux/reducers/embeddings/initialState';
-import initialCellSetsState from '../../../../../../../redux/reducers/cellSets/initialState';
-import initialGeneExpressionState, { initialExpressionState } from '../../../../../../../redux/reducers/genes/initialState';
-import initialExperimentState from '../../../../../../../redux/reducers/experimentSettings/initialState';
-import { initialPlotConfigStates } from '../../../../../../../redux/reducers/componentConfig/initialState';
+import ContinuousEmbeddingPlot from '../../../components/sharedPlots/ContinuousEmbeddingPlot';
+import { initialEmbeddingState } from '../../../redux/reducers/embeddings/initialState';
+import initialCellSetsState from '../../../redux/reducers/cellSets/initialState';
+import initialGeneExpressionState, { initialExpressionState } from '../../../redux/reducers/genes/initialState';
+import initialExperimentState from '../../../redux/reducers/experimentSettings/initialState';
+import { initialPlotConfigStates } from '../../../redux/reducers/componentConfig/initialState';
 
 jest.mock('localforage');
 const mockStore = configureStore([thunk]);
