+++ src/components/sharedPlots/ContinuousEmbeddingPlot.jsx
@@ -4,13 +4,13 @@ import PropTypes from 'prop-types';
 import { Spin } from 'antd';
 import { Vega } from 'react-vega';
 
-import PlatformError from '../../../../../../components/PlatformError';
-import { generateSpec, generateData } from '../../../../../../utils/plotSpecs/generateEmbeddingContinuousSpec';
-import { loadEmbedding } from '../../../../../../redux/actions/embedding';
-import { loadGeneExpression, loadPaginatedGeneProperties } from '../../../../../../redux/actions/genes';
-import { loadCellSets } from '../../../../../../redux/actions/cellSets';
-import { loadProcessingSettings } from '../../../../../../redux/actions/experimentSettings';
-import { updatePlotConfig } from '../../../../../../redux/actions/componentConfig/index';
+import PlatformError from '../PlatformError';
+import { generateSpec, generateData } from '../../utils/plotSpecs/generateEmbeddingContinuousSpec';
+import { loadEmbedding } from '../../redux/actions/embedding';
+import { loadGeneExpression, loadPaginatedGeneProperties } from '../../redux/actions/genes';
+import { loadCellSets } from '../../redux/actions/cellSets';
+import { loadProcessingSettings } from '../../redux/actions/experimentSettings';
+import { updatePlotConfig } from '../../redux/actions/componentConfig/index';
 
 const ContinuousEmbeddingPlot = (props) => {
   const { experimentId, config, plotUuid } = props;
