+++ src/pages/experiments/[experimentId]/data-processing/configure-embedding/components/EmbeddingPreview.jsx
@@ -12,8 +12,8 @@ import plot1Pic from '../../../../../../../static/media/plot9.png';
 import plot2Pic from '../../../../../../../static/media/plot10.png';
 import CalculationConfig from './CalculationConfig';
 
-import CategoricalEmbeddingPlot from './CategoricalEmbeddingPlot';
-import ContinuousEmbeddingPlot from './ContinuousEmbeddingPlot';
+import CategoricalEmbeddingPlot from '../../../../../../components/sharedPlots/CategoricalEmbeddingPlot';
+import ContinuousEmbeddingPlot from '../../../../../../components/sharedPlots/ContinuousEmbeddingPlot';
 
 import {
   updatePlotConfig,
